+++ abbr2words/languages/es.py
@@ -14,6 +14,7 @@ from abbr2words.core import (
     ExpansionVariant,
 )
 from abbr2words.language_data.initialisms import (
+    SPANISH_REVIEWED_INITIALISMS,
     TECHNICAL_INITIALISMS,
     register_reviewed_initialisms,
 )
@@ -416,6 +417,7 @@ class SpanishAbbreviationExpander(AbbreviationExpander):
             )
         )
 
+        register_reviewed_initialisms(self, SPANISH_REVIEWED_INITIALISMS)
         register_reviewed_initialisms(self, TECHNICAL_INITIALISMS)
 
         self.add_abbreviation(